requirements.txt
streamlit~=1.17.0
streamlit-authenticator~=0.2.1
PyYAML~=6.0
numpy~=1.22.4
requests~=2.27.1
beautifulsoup4~=4.11.1
nltk~=3.6.6
scipy~=1.7.3
scikit-learn~=1.1.2
validators~=0.20.0
Scrapy~=2.7.1
tldextract~=3.4.0
lxml~=4.9.1
newspaper3k~=0.2.8
tqdm~=4.62.3
pandas~=1.4.4